requirements.txt
aiofile==3.9.0
annotated-types==0.7.0
anyio==4.6.2.post1
bcrypt==4.2.1
caio==0.9.17
certifi==2024.8.30
click==8.1.7
colorama==0.4.6
Deprecated==1.2.15
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.5
fastapi-cli==0.0.5
fastapi-pagination==0.12.32
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.0
idna==3.10
Jinja2==3.1.4
limits==3.14.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
nodeenv==1.9.1
packaging==24.2
passlib==1.7.4
prisma==0.15.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pydantic==2.10.2
pydantic_core==2.27.1
Pygments==2.18.0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.19
pytz==2024.2
PyYAML==6.0.2
rich==13.9.4
rsa==4.9
shellingham==1.5.4
six==1.16.0
slowapi==0.1.9
sniffio==1.3.1
starlette==0.41.3
tomlkit==0.13.2
typer==0.14.0
typing_extensions==4.12.2
uvicorn==0.32.1
watchfiles==1.0.0
websockets==14.1
wrapt==1.17.0
